+++ RAG.py
@@ -11,20 +11,22 @@ from time import time
 from tiktoken_ext import openai_public
 import tiktoken_ext
 from Parameters import get_parameters
+import os
 
 class RAG:
 	def __init__(self, host, model):
+		self.llm_name=model
+		self.host=host
 		#Settings.embed_model = OllamaEmbedding(base_url=host, model_name=model)
 		Settings.embed_model = HuggingFaceEmbedding(model_name="BAAI/bge-small-en-v1.5")
-		options = get_parameters()
-		Settings.llm = Ollama(model=model, request_timeout=300.0, base_url=host, additional_kwargs=options)
 		self.index = None
 		self.token_counter = TokenCountingHandler(tokenizer=tiktoken.encoding_for_model("gpt-3.5-turbo").encode)
 		Settings.callback_manager = CallbackManager([self.token_counter])
-		self.update_settings()
+		self.update_settings(host, model)
 
-	def update_settings(self):
+	def update_settings(self, host, model):
 		options = get_parameters()
+		Settings.llm = Ollama(model=model, request_timeout=600, base_url=host, additional_kwargs=options)
 		Settings.chunk_size = settings.chunk_size
 		Settings.chunk_overlap = settings.chunk_overlap
 		Settings.similarity_top_k = settings.similarity_top_k
@@ -54,10 +56,14 @@ class RAG:
 			displayError(e)
 			setStatus("Failed to index.")
 
-	def loadFolder(self, folder, setStatus):
+	def loadFolder(self, path, setStatus):
 		try:
 			start = time()
-			documents = SimpleDirectoryReader(folder, recursive=True).load_data()
+			if isinstance(path, str):
+				documents = SimpleDirectoryReader(path, recursive=True).load_data()
+			else:
+				documents = SimpleDirectoryReader(input_files=path).load_data()
+
 			self.index = VectorStoreIndex.from_documents(documents)
 			message = f"Indexed folder into {len(documents)} chunks in {time()-start:0.2f} seconds."
 			displayInfo("Index", message)
@@ -68,7 +74,7 @@ class RAG:
 
 	def ask(self, question):
 		self.token_counter.reset_counts()
-		self.update_settings()
+		self.update_settings(self.host, self.llm_name)
 		node_postprocessors = [SimilarityPostprocessor(similarity_cutoff=settings.similarity_cutoff)]
 		query_engine = self.index.as_query_engine(similarity_top_k=settings.similarity_top_k, node_postprocessors = node_postprocessors, response_mode=settings.response_mode, streaming=True)
 		self.response = query_engine.query(question)