+++ changelog.md
@@ -1,5 +1,7 @@
 # Change log
 
+* Index file
+* Fixed bug when index gets reset when changing model
 * Token counter on the status bar for RAG.
 * Fixed response mode not working.
 * Output similarity score for each chunk.